Finalize spawning of rocks.
