+++ server/core/data/foreground/miscellaneous.js
@@ -6,6 +6,5 @@ export default [{
   id: 231,
   name: 'Golden Plaque',
   examine: 'Scripture written from the Gods years and years ago. Feels interesting.',
-  experience: 0,
   actions: ['push', 'examine'],
 }];
