+++ server/core/data/foreground/resources.js
@@ -18,6 +18,7 @@ export default [{
   resources: 'copper-ore',
   actions: presetActions(['pickaxe']),
   type: 'mine',
+  respawnIn: '2s',
 }, {
   id: 281,
   name: 'Tin Rocks',
@@ -27,4 +28,5 @@ export default [{
   resources: 'tin-ore',
   actions: presetActions(['pickaxe']),
   type: 'mine',
+  respawnIn: '5s',
 }];
